refactor: rename sbrc-vm to sirc-vm
