+++ sirc-vm/sirc-vm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "sbrc_vm"
+name = "sirc_vm"
 version = "0.1.0"
 edition = "2021"
