+++ sirc-vm/sirc-vm/src/lib.rs
