+++ sirc-vm/sirc-vm/src/utils/cpu_from_bus.rs
