+++ sirc-vm/sirc-vm/src/utils/lines.rs
