+++ sirc-vm/sirc-vm/src/utils/mod.rs
