+++ sirc-vm/sirc-vm/src/utils/path.rs
