+++ sirc-vm/sirc-vm/tests/debug_adapter/server_test.rs
@@ -1,4 +1,4 @@
-use sbrc_vm::debug_adapter::server::{format_instruction_ref, parse_instruction_ref};
+use sirc_vm::debug_adapter::server::{format_instruction_ref, parse_instruction_ref};
 
 // TODO: Make these unit tests
 // https://doc.rust-lang.org/book/ch11-03-test-organization.html#testing-private-functions
