+++ sirc-vm/toolchain/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 [dependencies]
 clap = { version = "4.5.4", features = ["derive"] }
 peripheral_cpu = { path = "../peripheral-cpu" }
-sbrc_vm = { path = "../sbrc-vm" }
+sirc_vm = { path = "../sirc-vm" }
 nom = "7.1.3"
 nom-supreme = "0.8.0"
 nom_locate = "4.2.0"
